Split other settings components into their own files
